Rebasing changes onto drive with distance
Handles the changes in master for pre2020 and 2020 drivebase.
